+++ my-tiny-spring/src/main/java/com/yrb/tinyioc/factory/BeanFactory.java
@@ -5,11 +5,11 @@ import com.yrb.tinyioc.BeanDefinition;
 /**
  * @author bjyangrubing
  * @createTime 2017/8/7 14:32
- * Description: 抽象工厂类
+ * Description: 工厂接口
  */
 public interface BeanFactory
 {
 	Object getBean(String name);
 
-	void registerBeanDefinition(String name, BeanDefinition beanDefinition);
+	void registerBeanDefinition(String name, BeanDefinition beanDefinition) throws Exception;
 }