+++ my-tiny-spring/src/test/java/com/yrb/tinyioc/BeanFactoryTest.java
@@ -13,16 +13,22 @@ public class BeanFactoryTest
 {
 
 	@Test
-	public void test(){
+	public void test() throws Exception
+	{
 
 		//1. 初始化beanFactory
 		BeanFactory beanFactory = new AutowireCapableBeanFactory();
 
-		//2. 注入bean
+		//2. bean的定义
 		BeanDefinition beanDefinition = new BeanDefinition();
-
 		beanDefinition.setBeanClassName("com.yrb.tinyioc.HelloWorldService");
 
+		//3. 设置属性
+		PropertyValues propertyValues = new PropertyValues();
+		propertyValues.addPropertyValue(new PropertyValue("text", "Hello, World"));
+		beanDefinition.setPropertyValues(propertyValues);
+
+		//4. 生成Bean
 		beanFactory.registerBeanDefinition("helloWorldService", beanDefinition);
 
 		//3. 获取bean